docs(emulator): add comprehensive hardware emulator documentation and fix compiler warnings, sprint 7
Add detailed documentation for the ESP32-based hardware emulator including
architecture overview, signal emulation specifications, web UI pages, API
endpoints, build configuration, and test scenarios.

Fix compiler warnings in emulator code:
- Correct printf format specifiers for unsigned types with proper casts
- Add __attribute__((unused)) to static const variables
- Improve JSON type checking using .is<T>() instead of containsKey()

+++ emulate_hardware/src/EmulatorWebServer.cpp
@@ -267,7 +267,8 @@ void EmulatorWebServer::handleGetSystemStatus(AsyncWebServerRequest* request) {
     uint32_t uptime = millis() / 1000;
     char uptimeStr[32];
     snprintf(uptimeStr, sizeof(uptimeStr), "%lud %02lu:%02lu:%02lu",
-             uptime / 86400, (uptime / 3600) % 24, (uptime / 60) % 60, uptime % 60);
+             (unsigned long)(uptime / 86400), (unsigned long)((uptime / 3600) % 24),
+             (unsigned long)((uptime / 60) % 60), (unsigned long)(uptime % 60));
     obj["uptime_formatted"] = uptimeStr;
 
     // WiFi info
